backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
httpx==0.27.2
pydantic==2.8.2

# Document processing (all free/open-source)
pdfplumber==0.11.4
pypdf==4.3.1
python-docx==1.1.2
python-pptx==1.0.2
openpyxl==3.1.5
Pillow==10.4.0

# OCR (optional - needs tesseract system package)
# pytesseract==0.3.13
